requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
pytesseract
Pillow
opencv-python-headless 